+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- possibility to change break/microbreak ideas via editing config file
+
 ## [0.5.1] - 2016-12-05
 ### Fixed
 - some grammar